change AutoCombineDust logic
As written, the AutoCombineDust() routine would combine as many inputs, as available, that match the set criteria. For example, if you set nAutoCombineThreshold to 200 and expect it to produce piles of up to 200 DMD from dust.. it won't do so. It will produce pile as big as you have inputs with values under 200 DMD each.

This change modifies the algorithm to combine dust in piles of size up to the setting.