Revert AutoCombineDust change
On a second though, the original logic of this routine is still useful - you set the values tnder which piles are considered "dust" and all these are collected in one larger pile. Beware, this might have strange consequences, such as not having enough value to pay for the transaction fees.